Move BoundedContext model
