+++ spec/Nidup/Architool/Infrastructure/Filesystem/FsBoundedContextRepositorySpec.php
@@ -2,8 +2,8 @@
 
 namespace spec\Nidup\Architool\Infrastructure\Filesystem;
 
-use Nidup\Architool\Domain\BoundedContext;
-use Nidup\Architool\Domain\BoundedContext\Layer;
+use Nidup\Architool\Domain\Model\BoundedContext;
+use Nidup\Architool\Domain\Model\BoundedContext\Layer;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Symfony\Component\Filesystem\Filesystem;
