+++ src/Nidup/Architool/Application/BoundedContext/CreateBoundedContextsHandler.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Nidup\Architool\Application\BoundedContext;
 
 use Nidup\Architool\Application\BoundedContext\CreateBoundedContexts;
-use Nidup\Architool\Domain\BoundedContext;
+use Nidup\Architool\Domain\Model\BoundedContext;
 use Nidup\Architool\Domain\BoundedContextRepository;
 
 final class CreateBoundedContextsHandler
